Move vendor-specific values to separate file
For reuse in the flexible-event directory.
